Adicionei fornecedor com regra de exclusao e padronizei retorno de produto com DTO

+++ src/main/java/com/gregory/controleestoque/controller/ProdutoController.java
@@ -6,8 +6,11 @@ import com.gregory.controleestoque.repository.ProdutoRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import com.gregory.controleestoque.dto.ProdutoResponseDTO;
+import com.gregory.controleestoque.model.Fornecedor;
 
 
 
@@ -24,58 +27,44 @@ public class ProdutoController {
         this.fornecedorRepository = fornecedorRepository;
     }
 
-    @GetMapping
-    public List<Produto> listar() {
-        return produtoRepository.findAll();
+    @GetMapping // lista os produtos
+    public List<ProdutoResponseDTO> listar() {
+        List<Produto> produtos = produtoRepository.findAll();
+        List<ProdutoResponseDTO> resposta = new ArrayList<>();
+
+        for(Produto produto : produtos) {
+            resposta.add(converterParaDTO(produto));
+        }
+        return resposta;
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<Produto> buscarPorId(@PathVariable Long id)  {
-        Optional<Produto> produto = produtoRepository.findById(id);
+    @GetMapping("/{id}") // retorna o produto para o front
+    public ResponseEntity<ProdutoResponseDTO> buscarPorId(@PathVariable Long id)  {
+        Optional<Produto> produtoOptional = produtoRepository.findById(id);
 
-        if (produto.isPresent()) {
-            return ResponseEntity.ok(produto.get());
+        if (produtoOptional.isPresent()) {
+            ProdutoResponseDTO dto = converterParaDTO(produtoOptional.get());
+            return ResponseEntity.ok(dto);
         }
-
         return ResponseEntity.notFound().build();
     }
 
-    @PostMapping
-    public Produto cadastrar(@RequestBody Produto produto) {
-        if (produto.getFornecedorId() == null) {
-            throw new RuntimeException("Fornecedor Obrigatorio");
-        }
-        if (!fornecedorRepository.existsById(produto.getFornecedorId())) {
-            throw new RuntimeException(("Fornecedor nao encontrado"));
-        }
-        return produtoRepository.save(produto);
+    @PostMapping // retorna um cadastro no padrao DTO para o front
+    public ProdutoResponseDTO cadastrar(@RequestBody Produto produto) {
+        validarProduto(produto);
+
+        Produto produtoSalvo = produtoRepository.save(produto);
+        return converterParaDTO(produtoSalvo);
     }
 
-    @PostMapping("/teste")
-    public Produto criarProdutoTeste() {
-        Produto produto = new Produto();
-        produto.setNome("Produto Teste");
-        produto.setDescricao("Produto basico para teste");
-        produto.setPreco(new BigDecimal("9.90"));
-        produto.setQuantidade(10);
-        produto.setEstoqueMinimo(2);
-        produto.setCategoria("Teste");
-        produto.setFornecedorId(1L);
-        return produtoRepository.save(produto);
-    }
 
-    @PutMapping("/{id}")
-    public Produto atualizar(@PathVariable Long id, @RequestBody Produto produtoAtualizado) {
-
-        if (produtoAtualizado.getFornecedorId() == null) {
-            throw new RuntimeException("Fornecedor Obrigatorio");
-        }
-        if (!fornecedorRepository.existsById(produtoAtualizado.getFornecedorId())) {
-            throw new RuntimeException(("Fornecedor nao encontrado"));
-        }
+    @PutMapping("/{id}") // retorna o atualizar no padrao DTO para o front
+    public ProdutoResponseDTO atualizar(@PathVariable Long id, @RequestBody Produto produtoAtualizado) {
+        validarProduto(produtoAtualizado);
 
         Produto produto = produtoRepository.findById(id)
-            .orElseThrow(() -> new RuntimeException("Produto nao encontrado"));
+                .orElseThrow(() -> new RuntimeException("Produto nao encontrado"));
+
 
         produto.setNome(produtoAtualizado.getNome());
         produto.setDescricao(produtoAtualizado.getDescricao());
@@ -85,12 +74,84 @@ public class ProdutoController {
         produto.setCategoria(produtoAtualizado.getCategoria());
         produto.setFornecedorId(produtoAtualizado.getFornecedorId());
 
-        return produtoRepository.save(produto);
+        Produto produtoSalvo = produtoRepository.save(produto);
+        return converterParaDTO(produtoSalvo);
     }
 
-    @DeleteMapping("/{id}")
-    public void deletar(@PathVariable Long id) {
-        produtoRepository.deleteById(id);
+    @DeleteMapping("/{id}") // Deletar produto
+    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+        Optional<Produto> produtoOptional = produtoRepository.findById(id);
+
+        if (produtoOptional.isPresent()) {
+            produtoRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
+
+    // meteodo para validar produto
+    private void validarProduto(Produto produto) {
+        if (produto.getFornecedorId() == null) {
+            throw new RuntimeException("Fornecedor Obrigatorio");
+        }
+
+        if (produto.getNome() == null || produto.getNome().trim().isEmpty()) {
+            throw new RuntimeException("Nome Obrigatorio");
+        }
+
+        if (!fornecedorRepository.existsById(produto.getFornecedorId())) {
+            throw new RuntimeException("Fornecedor nao encontrado");
+        }
+
+        if (produto.getPreco() == null) {
+            throw new RuntimeException("Preco Obrigatorio");
+        }
+
+        if (produto.getPreco().compareTo(BigDecimal.ZERO) < 0) {
+            throw new RuntimeException("Preco nao pode ser negativo");
+        }
+
+        if (produto.getQuantidade() < 0) {
+            throw new RuntimeException("Quantidade nao pode ser negativa");
+        }
+
+        if (produto.getEstoqueMinimo() < 0) {
+            throw new RuntimeException("Estoque minimo nao pode ser negativo");
+        }
+
+        if (produto.getCategoria() == null || produto.getCategoria().trim().isEmpty()) {
+            throw new RuntimeException("Categoria obrigatoria");
+        }
+    }
+
+    // esse metodo serve para responder o json melhor para o front
+    private ProdutoResponseDTO converterParaDTO(Produto produto) {
+        ProdutoResponseDTO dto = new ProdutoResponseDTO();
+
+        dto.setId(produto.getId());
+        dto.setNome(produto.getNome());
+        dto.setDescricao(produto.getDescricao());
+        dto.setPreco(produto.getPreco());
+        dto.setQuantidade(produto.getQuantidade());
+        dto.setEstoqueMinimo(produto.getEstoqueMinimo());
+        dto.setCategoria(produto.getCategoria());
+        dto.setFornecedorId(produto.getFornecedorId());
+
+        if (produto.getQuantidade() <= produto.getEstoqueMinimo()) {
+            dto.setStatusEstoque("Estoque Baixo");
+        } else {
+            dto.setStatusEstoque("Normal");
+        }
+
+        if (produto.getFornecedorId() != null) {
+            Optional<Fornecedor> fornecedorOptional = fornecedorRepository.findById(produto.getFornecedorId());
+
+            if (fornecedorOptional.isPresent()) {
+                dto.setNomeFornecedor(fornecedorOptional.get().getNome());
+            }
+        }
+       return dto;
+    }
+
 }
